+++ public/index.html
@@ -1923,5 +1923,5 @@ Here is a card with filled data</textarea
       </section>
     </main>
   </body>
-  <script src="../src/main.js"></script>
+  <script src="bundle.js"></script>
 </html>
